requirements.txt
arxiv
feedparser
google-genai
notion-client
httpx
python-dotenv
